+++ src/main/scala/org/constellation/ConstellationApp.scala
@@ -1,0 +1,5 @@
+package org.constellation
+
+class ConstellationApp(val constellationNode: ConstellationNode) {
+
+}
