+++ src/test/scala/org/constellation/app/ConstellationApp.scala
@@ -1,0 +1,16 @@
+package org.constellation.app
+import org.constellation.E2E
+
+class ConstellationApp extends E2E {
+
+  "ConstellationApp" should "successfully launch ConstellationNode" in {
+    assert(true)
+  }
+
+  "ConstellationApp" should "be able to deploy a state channel" in {
+    assert(true)
+  }
+  "ConstellationApp" should "be able to connect to deployed state channel" in {
+    assert(true)
+  }
+}
